+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Release Notes
 
 **Breaking Changes**
 
-- The improved TransactionObserverType protocol lets adopting types modify the database after a successful commit or rollback, and cancel a transaction with an error that pops up to the application code.
+- The improved TransactionObserverType protocol lets adopting types modify the database after a successful commit or rollback, and abort a transaction with an error.
 
 
 ## 0.19.0